+++ lib/Element.js
@@ -7,6 +7,7 @@ var NAMESPACE = utils.NAMESPACE;
 var attributes = require('./attributes');
 var Node = require('./Node');
 var NodeList = require('./NodeList');
+var NodeUtils = require('./NodeUtils');
 var FilteredElementList = require('./FilteredElementList');
 var DOMException = require('./DOMException');
 var DOMTokenList = require('./DOMTokenList');
@@ -98,7 +99,7 @@ Element.prototype = Object.create(ContainerNode.prototype, {
       // "the attribute must return the result of running the HTML fragment
       // serialization algorithm on a fictional node whose only child is
       // the context object"
-      return this._serializeOne({ nodeType: 0 });
+      return NodeUtils.serializeOne(this, { nodeType: 0 });
     },
     set: function(v) {
       var document = this.ownerDocument;
