+++ lib/Node.js
@@ -3,8 +3,8 @@ module.exports = Node;
 
 var EventTarget = require('./EventTarget');
 var LinkedList = require('./LinkedList');
+var NodeUtils = require('./NodeUtils');
 var utils = require('./utils');
-var NAMESPACE = utils.NAMESPACE;
 
 // All nodes have a nodeType and an ownerDocument.
 // Once inserted, they also have a parentNode.
@@ -37,45 +37,6 @@ var DOCUMENT_POSITION_FOLLOWING               = Node.DOCUMENT_POSITION_FOLLOWING
 var DOCUMENT_POSITION_CONTAINS                = Node.DOCUMENT_POSITION_CONTAINS = 0x08;
 var DOCUMENT_POSITION_CONTAINED_BY            = Node.DOCUMENT_POSITION_CONTAINED_BY = 0x10;
 var DOCUMENT_POSITION_IMPLEMENTATION_SPECIFIC = Node.DOCUMENT_POSITION_IMPLEMENTATION_SPECIFIC = 0x20;
-
-var hasRawContent = {
-  STYLE: true,
-  SCRIPT: true,
-  XMP: true,
-  IFRAME: true,
-  NOEMBED: true,
-  NOFRAMES: true,
-  PLAINTEXT: true
-};
-
-var emptyElements = {
-  area: true,
-  base: true,
-  basefont: true,
-  bgsound: true,
-  br: true,
-  col: true,
-  embed: true,
-  frame: true,
-  hr: true,
-  img: true,
-  input: true,
-  keygen: true,
-  link: true,
-  meta: true,
-  param: true,
-  source: true,
-  track: true,
-  wbr: true
-};
-
-var extraNewLine = {
-  /* Removed in https://github.com/whatwg/html/issues/944
-  pre: true,
-  textarea: true,
-  listing: true
-  */
-};
 
 Node.prototype = Object.create(EventTarget.prototype, {
 
@@ -732,75 +693,7 @@ Node.prototype = Object.create(EventTarget.prototype, {
   serialize: { value: function() {
     var s = '';
     for (var kid = this.firstChild; kid !== null; kid = kid.nextSibling) {
-      s += kid._serializeOne(this);
-    }
-    return s;
-  }},
-  _serializeOne: { value: function(parent) {
-    var kid = this, s = '';
-    switch(kid.nodeType) {
-      case 1: //ELEMENT_NODE
-        var ns = kid.namespaceURI;
-        var html = ns === NAMESPACE.HTML;
-        var tagname = (html || ns === NAMESPACE.SVG || ns === NAMESPACE.MATHML) ? kid.localName : kid.tagName;
-
-        s += '<' + tagname;
-
-        for(var j = 0, k = kid._numattrs; j < k; j++) {
-          var a = kid._attr(j);
-          s += ' ' + attrname(a);
-          if (a.value !== undefined) s += '="' + escapeAttr(a.value) + '"';
-        }
-        s += '>';
-
-        if (!(html && emptyElements[tagname])) {
-          var ss = kid.serialize();
-          if (html && extraNewLine[tagname] && ss.charAt(0)==='\n') s += '\n';
-          // Serialize children and add end tag for all others
-          s += ss;
-          s += '</' + tagname + '>';
-        }
-        break;
-      case 3: //TEXT_NODE
-      case 4: //CDATA_SECTION_NODE
-        var parenttag;
-        if (parent.nodeType === ELEMENT_NODE &&
-          parent.namespaceURI === NAMESPACE.HTML)
-          parenttag = parent.tagName;
-        else
-          parenttag = '';
-
-        if (hasRawContent[parenttag] ||
-            (parenttag==='NOSCRIPT' && parent.ownerDocument._scripting_enabled)) {
-          s += kid.data;
-        } else {
-          s += escape(kid.data);
-        }
-        break;
-      case 8: //COMMENT_NODE
-        s += '<!--' + kid.data + '-->';
-        break;
-      case 7: //PROCESSING_INSTRUCTION_NODE
-        s += '<?' + kid.target + ' ' + kid.data + '?>';
-        break;
-      case 10: //DOCUMENT_TYPE_NODE
-        s += '<!DOCTYPE ' + kid.name;
-
-        if (false) {
-          // Latest HTML serialization spec omits the public/system ID
-          if (kid.publicID) {
-            s += ' PUBLIC "' + kid.publicId + '"';
-          }
-
-          if (kid.systemId) {
-            s += ' "' + kid.systemId + '"';
-          }
-        }
-
-        s += '>';
-        break;
-      default:
-        utils.InvalidState();
+      s += NodeUtils.serializeOne(kid, this);
     }
     return s;
   }},
@@ -808,7 +701,7 @@ Node.prototype = Object.create(EventTarget.prototype, {
   // Non-standard, but often useful for debugging.
   outerHTML: {
     get: function() {
-      return this._serializeOne({ nodeType: 0 });
+      return NodeUtils.serializeOne(this, { nodeType: 0 });
     },
     set: utils.nyi,
   },
@@ -835,46 +728,3 @@ Node.prototype = Object.create(EventTarget.prototype, {
   DOCUMENT_POSITION_CONTAINED_BY: { value: DOCUMENT_POSITION_CONTAINED_BY },
   DOCUMENT_POSITION_IMPLEMENTATION_SPECIFIC: { value: DOCUMENT_POSITION_IMPLEMENTATION_SPECIFIC },
 });
-
-function escape(s) {
-  return s.replace(/[&<>\u00A0]/g, function(c) {
-    switch(c) {
-    case '&': return '&amp;';
-    case '<': return '&lt;';
-    case '>': return '&gt;';
-    case '\u00A0': return '&nbsp;';
-    }
-  });
-}
-
-function escapeAttr(s) {
-  var toEscape = /[&"\u00A0]/g;
-  if (!toEscape.test(s)) {
-      // nothing to do, fast path
-      return s;
-  } else {
-      return s.replace(toEscape, function(c) {
-        switch(c) {
-        case '&': return '&amp;';
-        case '"': return '&quot;';
-        case '\u00A0': return '&nbsp;';
-        }
-      });
-  }
-}
-
-function attrname(a) {
-  var ns = a.namespaceURI;
-  if (!ns)
-    return a.localName;
-  if (ns === NAMESPACE.XML)
-    return 'xml:' + a.localName;
-  if (ns === NAMESPACE.XLINK)
-    return 'xlink:' + a.localName;
-
-  if (ns === NAMESPACE.XMLNS) {
-    if (a.localName === 'xmlns') return 'xmlns';
-    else return 'xmlns:' + a.localName;
-  }
-  return a.name;
-}
